style: improve login form styling
