+++ crud/templates/login.html
@@ -1,31 +1,84 @@
-{% extends "base.html" %}
-{%block content%}
+{% block content %}
 
+<head>
+    {% load static %}
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE-edge"> 
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  
+    <title>Oficina de Filantropía</title>
+    <!--<link rel="stylesheet" type="text/css" href="{% static 'styles/base.css' %}">-->
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css">
+</head>
+    <section class="vh-100" style="background-color: #3255a4;">
+      <div class="container py-5 h-100 ">
+        <div class="row d-flex justify-content-center align-items-center h-100">
+          <div class="col col-xl-10">
+            <div class="card" style="border-radius: 1rem;">
+              <div class="row g-0">
+                <div class="col-md-6 col-lg-5 d-none d-md-block" style="background: url('{% static 'images/Foto_Icesi7.jpg' %}') no-repeat center center; border-radius: 1rem 0 0 1rem; background-size: cover;" >
+                </div>
 
-|<main class="container">
-    <div class="row">
-        <section class="col-md-4 offset-md-4">
-                <b>{{ title }}</b>
+                <div class="col-md-6 col-lg-7 d-flex align-items-center">
+                  <div class="card-body p-4 p-lg-5 text-black">
     
-                <br>
-                <br>
-                <form action="/login/" method = "POST" class="card card-body mt-5">
-                    <h1 class="text-center">Log in</h1>
-                    {% csrf_token %}
-                    <div class="mb-3">
-                        <label for="username">Nombre de usuario:</label>
-                        <input type="text" name="username" id="username"  class="form-control" placeholder="Escriba su nombre de usuario">
-                    </div>
+                    <form  class="needs-validation" novalidate action="/" method="POST" >
     
-                    <div class="mb-3">
-                        <label for="password1">Contraseña:</label>
-                        <input type="password" name="password1" id="password1"  class="form-control" placeholder="Escriba su contraseña">
-                    </div>
-                    <button class="btn btn-primary">Log In</button>
-                </form>
+                      <div class="d-flex align-items-center mb-3 pb-1">
+                        <img src="{% static 'images/logoIcesi.png' %}"  alt="login form" class="img-fluid" style="border-radius: 1rem 0 0 1rem;">
+                      </div>
+                      <h3 class="fw-normal mb-3 pb-3" style="letter-spacing: 1px;">Ingrese en su cuenta</h3>
+                      {% csrf_token %}
 
-        </section>
-    </div>
-</main>
+                      <div class="form-floating mb-3">
+                        <input type="text" name="username" id="username" class="form-control" placeholder="Nombre de Usuario" required>
+                        <label for="username">Nombre de usuario</label>
+                        <div class="invalid-feedback">
+                          Elija un nombre de usuario
+                        </div>
+                      </div>
+                      
 
-{% endblock %}
+                      <div class="form-floating mb-3">
+                        <input type="password" name="password" id="password"  class="form-control" placeholder="Contraseña" required>
+                        <label for="password">Contraseña:</label>
+                        <div class="invalid-feedback">
+                          Escriba su contraseña
+                        </div>
+                      </div>
+
+                      <p>{{error}}</p>
+
+                      <div class=" mt-4 d-flex justify-content-around align-items-center mb-4">
+                        <button class="btn btn-outline-primary btn-block">Ingresar</button>
+                      </div>
+                    </form>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </section>
+
+    <script>
+      // Example starter JavaScript for disabling form submissions if there are invalid fields
+      (() => {
+        'use strict'
+        // Fetch all the forms we want to apply custom Bootstrap validation styles to
+        const forms = document.querySelectorAll('.needs-validation')
+        // Loop over them and prevent submission
+        Array.from(forms).forEach(form => {
+          form.addEventListener('submit', event => {
+            if (!form.checkValidity()) {
+              event.preventDefault()
+              event.stopPropagation()
+            }
+            form.classList.add('was-validated')
+          }, false)
+        })})()
+    </script>
+{% endblock content %}
+
+
