+++ crud/views/login.py
@@ -1,0 +1,27 @@
+from django.shortcuts import render, redirect, get_object_or_404
+from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
+from django.contrib.auth import login, logout, authenticate
+from django.contrib.auth.models import User
+from django.db import IntegrityError
+from django.utils import timezone
+from django.contrib.auth.decorators import login_required
+
+
+@login_required
+def home(request):
+    return render(request, 'home.html')
+
+
+def signin(request):
+    if request.method == 'GET':
+        return render(request, 'login.html', {"form": AuthenticationForm})
+    else:
+        user = authenticate(
+            request, username=request.POST['username'], password=request.POST['password'])
+        if user is None:
+            # Authentication failed
+            return render(request, 'login.html', {"form": AuthenticationForm, "error": "Usuario o contraseña es incorrecto."})
+
+        # Authentication succeeded
+        login(request, user)
+        return redirect('home')
