+++ crud/views/views.py
@@ -9,9 +9,6 @@ from crud.models import Student
 
 
 # Create your views here.
-def home(request):
-    return render(request, 'home.html')
-
 
 def signUp(request):
     if request.method == 'GET':
@@ -31,33 +28,18 @@ def signUp(request):
 
             return HttpResponse('Contrasenas incorrectas')
 
-          
-def logIn(request):
-    if request.method == 'GET':
-        return render(request, 'login.html', {
-            'form': AuthenticationForm
-        })
-    else:
-        user = authenticate(
-            request, username=request.POST['username'], password=request.POST['password'])
-        if user is None:
-            return HttpResponse('No existe el usuario')
-        else:
-            login(request, user)
-            return redirect("/postlog")
-
 
 @login_required
 def postLog(request):
     return render(request, 'logout.html')
 
 
-@login_required
-def singout(request):
-    logout(request)
-    return redirect(home)
+# @login_required
+# def singout(request):
+#    logout(request)
+#    return redirect(home)
 
-  
+
 def confirmacion(request):
     return render(request, 'students_confirm.html')
